header component start ed, Input component started, Colors in constants folder, moved NewAppScreen contents into Start.js so it is still accessible
