+++ constants/Colors.js
@@ -1,0 +1,4 @@
+export default {
+  primary: 'blue',
+  secondary: '#e3e7ed',
+};
